+++ FinanceTrack/Core/Services/IncomesRepository.swift
@@ -4,6 +4,6 @@ protocol IncomesRepository {
     func listAll() -> [Income]
     func listAll(period: Period) -> [Income]
     
-    func add(amount: Int, date: Date)
+    func add(amount: Double, date: Date)
     func getTotal(for period: Period) -> Double
 }
